+++ rust/oas-common/src/types/post.rs
@@ -10,13 +10,14 @@ use serde::{Deserialize, Serialize};
 use serde_json::json;
 
 use super::Media;
+// use super::Feed;
 
 #[derive(Serialize, Deserialize, Debug, Clone, Default, JsonSchema)]
 #[serde(rename_all = "camelCase")]
 pub struct Post {
     pub identifier: Option<String>,
     pub headline: Option<String>,
-    pub post_abstract: Option<String>,
+    pub r#abstract: Option<String>,
     pub contributor: Option<String>,
     pub url: Option<String>,
     pub date_published: Option<String>,
@@ -31,7 +32,8 @@ pub struct Post {
     pub creator: Vec<String>,
     #[serde(default)]
     pub media: Vec<Reference<Media>>,
-
+    // #[serde(default)]
+    // pub feed: Reference<Feed>,
     pub transcript: Option<String>,
 
     #[serde(flatten)]
